Set File Uri in DB
The File Uri is usually updated only after the
download is completed on receiver end, due to which
we are unaware of the partial downloaded content's path
while trying to resume a paused file transfer.
Fixed the same by storing the file URI as soon as it is
known at the beginning of the download trigger. A client
can anyway know that that file Uri does not refer to a
completely downloaded file until the state has reached
TRANSFERRED and/or when progress has reached 100%. And
then the resume download works again as it should.