CR025 Changed "Joyn" name to "Rcs"
Almost all the existence of the word "Joyn" in either method names,
parameters, options, Exceptions, Classes, Interfaces etc are
replaced in the entire Terminal Api with the word "Rcs".
This applies to all occurrences of the word "Joyn" wherever
it is present in the Terminal Api definitions currently.
Exceptions are network payload constants and some other externally
shared constants.

Change-Id: I832ae6ffd4c35c05d2ed7ea35c3908ace6e50ab9