Use Secure SIP if TLS Protocol
During DNS queries if NAPTR records are not sent by the server or NAPTR query from the device fails, then a query for direct SRV lookup is generated with the ims proxy address.

While generating this query _sips. should be used, If the underlying protocol is TLS.

Reference: RFC3261, RFC3329 and as specified in 3GPP TS 24.229