Synchronize the DB attach/detach methods
There was a possibility for race condition to attach to database
with same alias causing a SQLite exception.

Solve this by synchronized statements that surrounds the method that
attach/detach to DB and also sets a flag stating the status of the same.

Testing:
arcsine test that caught the error

Change-Id: I9e09a35f65d229cd10ef53455b4c7b23318ff484