Auto reject 1-1 invitation which doesn't contain first message
Invitation will be rejected if it is OMA SIMPLE IM solution
but it doesn't contains first message.

Reference to spec: Rich Communication Suite 5.1 Advanced
Communications Services and Client Specification Version 4.0
Page 187 3.3.4.2 Technical Realization of 1-to-1 Chat features
when using OMA SIMPLE IM At the technical level the 1-to-1 Chat
service implemented using OMA SIMPLE IM extends the concepts
described in section 3.3.4.1 with the following concepts: For
OMA SIMPLE IM, first message is always included in a CPIM/IMDN
wrapper carried in the SIP INVITE request. So the configuration
parameter FIRST MSG IN INVITE defined in Table 80 is always set
to 1.

Change-Id: I67c867d1b8d1498f8beee2b0405263199e6ee265