Update default IsComposingTimeout value to 15s
Update default IsComposingTimeout value to 15s from 5s according
to RFC3994.

"If the user stops composing for more than a configured time
interval, the idle timeout, the state transitions to "idle", and
an "idle" status message is sent.  If the user starts composing
again while in "idle" state, the state transitions to "active",
and the corresponding status message is sent.  Unless otherwise
configured by the user, the idle timeout SHOULD have a default
value of 15 seconds."

Change-Id: I72d9851910167ac54d8ef351ad6bdbad212ce87b