Fix "States arrive in wrong order to ImageSharingListner" issue
The root cause of the issue is startMediaSession() carries out the
actually data transfer process and it is blocking the thread. So
the state is set to STARTED after the transfer is done.

startMediaSession() was running on a new thread before and it was
changed in exception handling pacth.

The solution is to set state to STARTED after send out ACK and then
carry out the data transfer.

Also changed function name from startMediaSession() to
startMediaTransfer(), new name explains clearer of what the function
is doing.

Change-Id: I1e50b5ead4f7e191255bff1f8149e939a7e9fcfb