Allow only secure connection for file upload
According to Rich Communication Suite 5.1 Advanced
Communications Services and Client Specification Version 4.0
3.5.4.8.3 File transfer procedure
3.5.4.8.3.1 Sender procedures
This specification uses the term “HTTP POST” and “HTTP GET”
as a generic reference to the action of using the POST or GET
method. However, it is strongly recommended that whenever the
POST action contains sensitive information such as a user ID
or password, the action should take place over a secure
connection and/or via HTTPS explicitly.

Change-Id: Iefb3fc59ffbe4ea60af8f8f7684d841e2b8b0fd3