Fix setting state as ABORTED after receiving SIP BYE
In Incoming MSRP FileSharingSession, SIP BYE can be
received from the sender if either the sender wishes to
abort the session or when the transfer of file is
completed. In both cases, we need to close the session
and perform the clean up activities of the session. We
will broadcast the state as ABORTED and ABORTED_BY_REMOTE
only if the file was not received successfully.

Change-Id: Id7f24dd0c903a9830b0805a05300dc10250dd96e