Clean up and exception handling for Utils
Updated Util classes to properly handle exceptions and
removed instance of generic Exception class.

Also did a cleanup of some util classes which were not
doing anything specific activity like :
- AppUtils.java
- HttpUtils.java

Moved AppUtils getApplicationVersion to TerminalInfo
so to have only one file that take care of all
information related to terminal.

Note : Exception handling for XdmManager.java is
still done partially to suffice this commit, Cleanup
for the whole presence package is kept as a task
in backlog and will done while implementing PRESENCE
feature itself as during feature implementation we can
also test the changes we do which at this moment is not
quite possible.

Change-Id: I843ae8ecbc9afe8dad25a8712ca18c0952904065