S&F notification sessions are always auto-accepted
According to specification :
3.3.4.1.5 Delivering stored disposition notifications :
To be able to deliver delivered/displayed notifications
that were stored to a sender's device that has come online
again, without disrupting the user experience, the Messaging
Server supporting the store and forward functionality shall
initiate a special session for the purpose of delivering these
notifications. This special session shall be automatically
accepted by the device.

Change-Id: Icea90f2afb364aa047963021692f636e684dbea5
Signed-off-by: Kristoffer Karlsson <kristoffer.karlsson2@sonymobile.com>