Exception handling updates and removal of SipException instances
Made changes to remove instances where the generic Exception
was being catched, as part of exception handling.

Refactor SipException instances so that we can distinguish them
as Payload & Netowrk exceptions.

Payload exceptions : That are real bugs and need to be corrected.
Network exceptions : That are temporary and mostly occur due to network
loss.

Change-Id: I00a946aa2b36d26bca9812eb910fda6fbea6aa8c