Automatically reject chat invitation if first message is empty
When receiving a chat invitation, if first message in the invite
is empty or doesn't exist, the invitation should be rejected
automaticallt.

Specification reference:

Rich Communication Suite 5.1 Advanced Communications Services
and Client Specification

Version 3.0

Page 182

3.3.4.2 Technical Realization of 1-to-1 Chat features when using
OMA SIMPLE IM

For OMA SIMPLE IM, first message is always included in a CPIM/IMDN
wrapper carried in the SIP INVITE request. So the configuration
parameter FIRST MSG IN INVITE defined in Table 77 is always set to
1. A client should always include "positive-delivery" in the value
for the Disposition-Notification header field in that message. That
means that the value of the header field is either "positive-delivery"
or "positive-delivery,display" depending on whether display
notifications were requested. The value of "negative-delivery" is not
used in RCS for 1-to-1 Chat. SIP INVITE requests for a one-to-one
session that carry a message in CPIM/IMDN wrapper shall be rejected
by the server unless they carry a Disposition-Notification header
that at least includes "positive-delivery".

Change-Id: I6d6ccf3f0ad0c6a59794fa93dc23176085dd6081
Signed-off-by: Kristoffer Karlsson <kristoffer.karlsson2@sonymobile.com>