Remove manually setting dialog path
Remove manually setting dialog path for TerminatingHttpFileSharing
session as this tries to send SIP BYE in this dialog path when this
FileSharingSession is getting closed which is not correct because
this type of session originally does not have a dialog path.
Also it is noticed that the dialog path is currently set only for
those TerminatingHttpFileSharingSession(s) that are created in
manual acceptance use-case.

Change-Id: Iadec1d8b36f76d464ad6961749475858ba0303e8
Signed-off-by: Kristoffer Karlsson <kristoffer.karlsson2@sonymobile.com>