ims_connection: workaround for white small icon in notification bar
With Android 5.0 Lollipop it is no longer possible to use colored icons in the notification area.
Only large icon supports colors but only small icon can be shown in notification bar.
Large icon are displayed in the drawer in addition with small icon then.
This is a temporary workaround waiting for small white icons to distinguish IMS connection from disconnection.