Bug correction for the Backup and restore feature
There has been a regression introduced while implementing exception
handling for Backup & Restore feature, which has been corrected
in this commit.

With these changes in cases where we fail to do Backup or Restore
operation, we clean up all the partial files that we copied till
that point as there should never be a partial success
in case of a failure.

Change-Id: Idbaeaa4879f61e199fa8e6f7073868722e139791
Signed-off-by: Kristoffer Karlsson <kristoffer.karlsson2@sonymobile.com>