Full stack trace not to be printed for exceptions not caused by bugs
Network and IOExceptions that are caused due to network loss
are considered as non-bugs reason being that there is no way
that we can handle such use-cases.
Hence for such cases we should not add full stack trace and it's
enough to just log there occurence as a notion.

Made such changes through this commit, also in some cases we were
logging twice (full trace & also in handleError) made changes in
such places also to log only once so as to avoid log spamming.

Change-Id: Ieb52e2c84f0494be582a2c5b91ea28bd26816e6a