Replaced String concat with StringBuilder append
While logging for exceptions in some places we
used .concat to append different id's so that we
can make the exception message more intutive for
debugging purposes, However in some cases these
id's are NULL and hence .concat ends up in
throwing Null-Pointer.
As it's not good that a exception message inturn
throws another exception as in such cases the
original trace is lost, Made changes in all such
places to use StringBuilder.append instead.

Changes are done only for the cases related to
Exception Handling.

Change-Id: I81f7a210fd7ce3d4651cf3b1cf18485d66318727