getFileIcon() incorrectly returns the uri to the file and not the icon
Upon using the file transfer api it seems that
IFileTransfer.getFileIcon() incorrectly returns the uri to the file
and not the icon. This is in particular apparent on the receiving side,
since he will never have a file icon for the incoming file transfer.
Fixed the same through this commit.