File icons hidden in a marked folder
File icons are shown in the album since the media scanner
finds them and categorises them as images.
The file icons are now placed in a new folder "rcs/fileicons/"
with a file named ".nomedia" so that the media scanner will
not scan the file icons folder and will not create preview
images in photo viewer app. This new folder is created if not
already existing at all times when the RCS core service is started.

Issue#185