Removed unsupported cipher-suite used for SIP TLS on Android M cookie
With Android M (6.0) below cipher suites are not supported anymore,
and keeping them caused IllegalArgumentException when request for
registration using SIP TLS on Vodafone network.(Wi-Fi reg. use-case).

In general also this will be an issue for any operator using SIP TLS
for registration.

Removed/Commented out suite entries:
- TLS_DH_anon_WITH_AES_128_CBC_SHA
- SSL_DH_anon_WITH_3DES_EDE_CBC_SHA

Instead of them the default suite
- TLS_RSA_WITH_AES_128_CBC_SHA
will be used.

This is also in alignment to the recommendation from RFC3261.

Change-Id: I12199a0c3686e24f12737414f5ecc9d6282fa9e1