Issue with ims services background handler instantiation
To run heavy operations on background IMS services
instantiate background handlers, however the handler
threads created were never stopped when the rcs service
is stopped, This has cause a leak as everytime we restart the
stack a new handler thread gets created, without removing the
old one.

If we start/stop the core stack quite often this can lead to
resource starvation and also possible leaks.

This commit proposes a fix for the same.

Change-Id: I573d05eeb2246f521b0b6fb24859cc635286baff
Fix: #Issue-179