De-register not performed when battery reaches critical threshold
When battery reaches to a critical level a controlled
de-registration should have to be performed and this should
result in stopping Ims Connection with
TerminationReason: TERMINATION_BY_BATTERY_LOW
and a SIP Register with Expire=0 (de-register) should been sent
to ims.
This flow was not properly implemented, Fixed the same through
this commit.

Change-Id: Icef683fb2a1669e3eb9d9e0ace8e1fbebf611590